Set file system permissions for addons in Docker image
If the directories/files containing the KKP addons are not readable by
group/others (e.g. because the user has a restrictive default umask)
when building the Docker image, the kubermatic-seed-controller is not
able to get the addon manifests. Therefore the file system permissions
will be set for the addon directories/files.

Signed-off-by: Daniel Kraus <daniel.kraus@kubermatic.com>